fix 解决 id 类型问题

+++ internal/miniblog/biz/article/article.go
@@ -17,10 +17,10 @@ import (
 type IArticleBiz interface {
 	Create(ctx context.Context, r *v1.CreateArticleRequest) (*v1.ArticleInfoResponse, error)
 	Update(ctx context.Context, r *v1.UpdateArticleRequest) (*v1.ArticleInfoResponse, error)
-	Publish(ctx context.Context, articleId int) error
-	Unpublish(ctx context.Context, articleId int) error
+	Publish(ctx context.Context, articleId uint64) error
+	Unpublish(ctx context.Context, articleId uint64) error
 	GetList(ctx context.Context, r *v1.ArticleListRequest) (*v1.GetArticleListResponse, error)
-	GetOne(ctx context.Context, id int) (*v1.GetArticleResponse, error)
+	GetOne(ctx context.Context, id uint64) (*v1.GetArticleResponse, error)
 }
 
 // articleBiz 文章业务实现
@@ -110,7 +110,7 @@ func (b *articleBiz) Update(ctx context.Context, r *v1.UpdateArticleRequest) (*v
 }
 
 // Publish 发布文章
-func (b *articleBiz) Publish(ctx context.Context, articleId int) error {
+func (b *articleBiz) Publish(ctx context.Context, articleId uint64) error {
 	article, err := b.ds.Articles().GetOne(articleId)
 	if err != nil {
 		return errno.ErrArticleNotFound
@@ -126,7 +126,7 @@ func (b *articleBiz) Publish(ctx context.Context, articleId int) error {
 }
 
 // Unpublish 下架文章
-func (b *articleBiz) Unpublish(ctx context.Context, articleId int) error {
+func (b *articleBiz) Unpublish(ctx context.Context, articleId uint64) error {
 	article, err := b.ds.Articles().GetOne(articleId)
 	if err != nil {
 		return errno.ErrArticleNotFound
@@ -191,7 +191,7 @@ func (b *articleBiz) GetList(ctx context.Context, r *v1.ArticleListRequest) (*v1
 }
 
 // GetOne 获取文章详情
-func (b *articleBiz) GetOne(ctx context.Context, id int) (*v1.GetArticleResponse, error) {
+func (b *articleBiz) GetOne(ctx context.Context, id uint64) (*v1.GetArticleResponse, error) {
 	article, err := b.ds.Articles().GetOne(id)
 	if err != nil {
 		return nil, err
